+++ R/removeConstantTerms.R
@@ -80,7 +80,7 @@ removeConstantTerms = function(formula, data){
 		fterms_new = fterms
 		for( x in excludeVar){
 
-			# fterms_new = gsub(x, '1', fterms_new)
+			# old: fterms_new = gsub(x, '1', fterms_new)
 
 			# make sure string goes to end, or end followed by )
 			fterms_new = gsub(paste0(x, '$'), '1', fterms_new)
@@ -89,7 +89,7 @@ removeConstantTerms = function(formula, data){
 
 		fterms_new = unique(fterms_new)
 
-		# fterms_new = array(sapply(excludeVar, function(x) gsub(x, '1', fterms)))
+		# old: fterms_new = array(sapply(excludeVar, function(x) gsub(x, '1', fterms)))
 		fterms_new = fterms_new[fterms_new!=""]
 	}else{
 		fterms_new = fterms